requirements.txt
anthropic==0.34.0
fastapi_poe==0.0.47
ln_markets==2.0.5
openai==1.41.0
pandas==2.0.3
python-dotenv==1.0.1
pytz==2021.3
yfinance==0.2.37
